fix: 代码审查修复(P0/P1/P2)与单测补齐(07审查报告)
P0 高危:
- offline 超时检查移到 receive_loop 顶部无条件执行(bms/mppt/dcdc 三节点),
  修复被无关帧"饿死"导致失联检测失效; 抽取 offline_utils 统一离线兜底判定
- FlightStatus.battery_remaining 契约统一为 0~100(跨 msg/节点/测试)
- bms_node 配置改 can1(默认, 匹配实际设备), 修正离线
- link_node udp_port 默认值 14551 改 20000, 避开数传映射段并校验
- rpi5 配置 can0/can1 角色修正(msgs 对应)
- io-check.sh 改只读探测波特率, 不再破坏性重配 CAN
- DCDC 温度 temp_with_offset 改 uint8_t, 修复 raw>127 过温被截断

P1 中危&源码:
- 各节点/协议/单测同步修正(见 07_审查报告 §3)
- 新抽 fc_alert_logic 规则引擎、offline_utils、test_mqtt_client 等单测
- 消息定义 BmsStatus/DcdcStatus/MpptStatus/LoRaSample 字段审计
- link/cloud 订阅改 SensorDataQoS(只取最新帧)降低背压内存

构建配置:
- 各包 CMakeLists/package.xml 补单测与依赖, 新增 .clang-tidy 检查配置

+++ src/airship_dcdc/src/dcdc_protocol.cpp
@@ -73,8 +73,8 @@ void parse_analog(const uint8_t * data, uint32_t len, DcdcData & out)
   out.input_voltage = scale_u16(get_u16_le(data, 0), 1.0f);
   out.output_voltage = scale_u16(get_u16_le(data, 2), 0.1f);
   out.output_current = scale_u16(get_u16_le(data, 4), 0.1f);
-  out.ambient_temp = temp_with_offset(static_cast<int8_t>(data[6]));
-  out.heatsink_temp = temp_with_offset(static_cast<int8_t>(data[7]));
+  out.ambient_temp = temp_with_offset(data[6]);
+  out.heatsink_temp = temp_with_offset(data[7]);
 }
 
 }  // namespace airship_dcdc
